+++ src/main/java/com/fellows/employee/dao/EmployeeDao.java
@@ -1,8 +1,8 @@
-package com.fellows.day6.dao;
+package com.fellows.employee.dao;
 
 import java.util.List;
 
-import com.fellows.day6.model.Employee;
+import com.fellows.employee.model.Employee;
 
 public interface EmployeeDao {
 	public List<Employee> getAllEmployees();
