+++ src/main/java/com/fellows/employee/service/EmployeeService.java
@@ -1,8 +1,8 @@
-package com.fellows.day6.service;
+package com.fellows.employee.service;
 
 import java.util.List;
 
-import com.fellows.day6.model.Employee;
+import com.fellows.employee.model.Employee;
 
 public interface EmployeeService {
 	  public List<Employee> getAllEmployees();
