+++ src/main/java/com/fellows/employee/service/EmployeeServiceImpl.java
@@ -1,12 +1,12 @@
-package com.fellows.day6.service;
+package com.fellows.employee.service;
 
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.fellows.day6.dao.EmployeeDao;
-import com.fellows.day6.model.Employee;
+import com.fellows.employee.dao.EmployeeDao;
+import com.fellows.employee.model.Employee;
 
 @Service
 public class EmployeeServiceImpl implements EmployeeService {
